+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added a test case for `terra-list` to support drag and drop in `terra-modal-manager`.
+  * Added aria-label to announce the visual label.
 
 ## 1.52.0 - (November 21, 2023)