+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added
+  * Added condition to announce the placeholder across all browsers.
   * Added visual focus dashed border for `terra-form-select` list options.
 
 * Fixed